Fixes of markups and adding tabs
